2025 day 9: 3.6 seconds on Raspberry Pi 5, but not a general solution
XXX my "part 2" code uses edge segments as a surrogate for blank
regions. It assumes that no parallel green edges abut without a gap.
It might also fail to exclude some rectangles in "concavities" that
are almost entirely outside the green region.

But it seems to work for my input file.